Gym/blogging/requirements.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
idna==2.10
jsmin==2.2.2
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
